+++ Logos/src/utopia/logos/database/LogosContext.scala
@@ -1,6 +1,7 @@
 package utopia.logos.database
 
 import utopia.flow.parse.json.{JsonParser, JsonReader}
+import utopia.flow.util.logging.{Logger, SysErrLogger}
 import utopia.vault.context.{VaultContext, VaultContextWrapper}
 import utopia.vault.database.{ConnectionPool, Tables}
 
@@ -17,10 +18,15 @@ object LogosContext extends VaultContextWrapper
 	
 	private var vaultContext: Option[VaultContext] = None
 	private var _jsonParser: JsonParser = JsonReader
+	private var _log: Logger = SysErrLogger
 	
 	
 	// COMPUTED -------------------------
 	
+	/**
+	 * @return The logging implementation used within this project
+	 */
+	implicit def log: Logger = _log
 	/**
 	  * @return The JSON processor used in this project
 	  */
@@ -37,9 +43,12 @@ object LogosContext extends VaultContextWrapper
 	
 	/**
 	  * @param vaultContext A context that provides the database interaction properties for this interface
+	 * @param jsonParser The JSON-parsing implementation used in this project
+	 * @param log Logging implementation used in this project. Default = write to System.err
 	  */
-	def setup(vaultContext: VaultContext, jsonParser: JsonParser) = {
+	def setup(vaultContext: VaultContext, jsonParser: JsonParser, log: Logger) = {
 		this.vaultContext = Some(vaultContext)
+		_log = log
 		_jsonParser = jsonParser
 	}
 	/**
@@ -48,8 +57,9 @@ object LogosContext extends VaultContextWrapper
 	  * @param databaseName Name of the database that contains the Logos tables
 	  * @param tables The interface for accessing the tables within this project
 	  * @param jsonParser The JSON-parsing implementation used in this project
+	  * @param log Logging implementation used in this project. Default = write to System.err
 	  */
 	def setup(exc: ExecutionContext, cPool: ConnectionPool, databaseName: String, tables: Tables,
-	          jsonParser: JsonParser): Unit =
-		setup(VaultContext(exc, cPool, databaseName, tables), jsonParser)
+	          jsonParser: JsonParser, log: Logger = SysErrLogger): Unit =
+		setup(VaultContext(exc, cPool, databaseName, tables), jsonParser, log)
 }