+++ Logos/src/utopia/logos/database/access/many/url/domain/DbDomains.scala
@@ -1,7 +1,6 @@
 package utopia.logos.database.access.many.url.domain
 
-import utopia.flow.collection.immutable.{Empty, Pair}
-import utopia.flow.operator.enumeration.End.First
+import utopia.logos.database.CachingVolatileMapStore
 import utopia.logos.model.partial.url.DomainData
 import utopia.logos.model.stored.url.Domain
 import utopia.vault.database.Connection
@@ -12,65 +11,17 @@ import utopia.vault.nosql.view.{UnconditionalView, ViewManyByIntIds}
   * @author Mikko Hilpinen
   * @since 20.03.2024, v0.2
   */
-object DbDomains extends ManyDomainsAccess with UnconditionalView with ViewManyByIntIds[ManyDomainsAccess]
+object DbDomains
+	extends CachingVolatileMapStore[String, String, Domain] with ManyDomainsAccess with UnconditionalView
+		with ViewManyByIntIds[ManyDomainsAccess]
 {
-	// OTHER	--------------------
+	override protected def standardize(value: String): String = value.toLowerCase
+	override protected def diff(proposed: Set[String], existing: Set[String]): Set[String] =
+		proposed.filterNot { d => existing.contains(d.toLowerCase) }
 	
-	/**
-	  * Stores the specified domains in the DB. Avoids inserting duplicates.
-	  * @param domains Domains to store
-	  * @param connection Implicit DB connection
-	  * @return Inserted domains (first) and existing domain entries (second).
-	  * Contains separate groups for inserted items (first) and items that already existed in the DB (second)
-	  */
-	def store(domains: Set[String])(implicit connection: Connection) = {
-		if (domains.nonEmpty) {
-			val lowerCaseDomains = domains.map { _.toLowerCase }
-			// Checks for existing entries
-			val existing = matching(lowerCaseDomains).pull
-			val existingDomainNames = existing.map { _.url.toLowerCase }.toSet
-			// Inserts missing entries
-			val inserted = model.insert(
-				domains.filterNot { url => existingDomainNames.contains(url.toLowerCase) }
-					.toVector.map { DomainData(_) }
-			)
-			
-			// Returns the values in two groups: Inserted & existing
-			Pair(inserted, existing)
-		}
-		else
-			Pair.twice(Empty)
-	}
-	
-	/**
-	 * Stores the specified domains in the DB. Avoids inserting duplicates.
-	 * @param values Values from which domains are extracted
-	 * @param extractDomain A function that extracts a domain from a value
-	 * @param merge A function that merges 3 values:
-	 *                  1. The inserted domain
-	 *                  1. The original value
-	 *                  1. Whether the domain was inserted
-	 * @param connection Implicit DB connection
-	 * @tparam A Type of original values
-	 * @tparam R Type of merge results
-	 * @return Merge results
-	 */
-	def storeFrom[A, R](values: Iterable[A])(extractDomain: A => String)(merge: (Domain, A, Boolean) => R)
-	                   (implicit connection: Connection) =
-	{
-		// Extracts the domain information
-		val valuesWithDomains = values.map { v => v -> extractDomain(v) }
-		// Stores the distinct domains and forms a map of the inserted data
-		val domainMap = store(valuesWithDomains.view.map { _._2 }.toSet).zipWithSide
-			.flatMap { case (domains, side) =>
-				val wasInserted = side == First
-				domains.map { d => d.url.toLowerCase -> (d, wasInserted) }
-			}
-			.toMap
-		// Merges the inserted data with the original data
-		valuesWithDomains.flatMap { case (value, domainStr) =>
-			domainMap.get(domainStr.toLowerCase).map { case (domain, wasInserted) => merge(domain, value, wasInserted) }
-		}
-	}
+	override protected def pullMatchMap(values: Set[String])(implicit connection: Connection): Map[String, Domain] =
+		matching(values).toMapBy { _.url.toLowerCase }
+	override protected def insertAndMap(values: Seq[String])(implicit connection: Connection): Map[String, Domain] =
+		model.insert(values.sorted.map { DomainData(_) }).view.map { d => d.url.toLowerCase -> d }.toMap
 }