+++ Logos/src/utopia/logos/database/access/many/url/path/DbRequestPaths.scala
@@ -1,18 +1,16 @@
 package utopia.logos.database.access.many.url.path
 
 import utopia.flow.collection.CollectionExtensions._
-import utopia.flow.collection.immutable.{Empty, OptimizedIndexedSeq, Pair}
-import utopia.flow.operator.Identity
-import utopia.flow.operator.enumeration.End.First
-import utopia.flow.operator.equality.EqualsExtensions._
+import utopia.flow.collection.immutable.Empty
+import utopia.flow.util.logging.Logger
+import utopia.logos.database.CachingVolatileMapStore
 import utopia.logos.database.access.many.url.domain.DbDomains
+import utopia.logos.model.cached.Link
 import utopia.logos.model.combined.url.DetailedRequestPath
 import utopia.logos.model.partial.url.RequestPathData
-import utopia.logos.model.stored.url.{Domain, RequestPath}
+import utopia.logos.model.stored.url.Domain
 import utopia.vault.database.Connection
 import utopia.vault.nosql.view.{UnconditionalView, ViewManyByIntIds}
-
-import scala.collection.View
 
 /**
   * The root access point when targeting multiple request paths at a time
@@ -22,35 +20,25 @@ import scala.collection.View
 object DbRequestPaths 
 	extends ManyRequestPathsAccess with UnconditionalView with ViewManyByIntIds[ManyRequestPathsAccess]
 {
-	/**
-	  * Stores the specified request paths to the database.
-	  * Avoids inserting duplicates
-	  * @param paths Paths to store
-	  * @param connection Implicit DB connection
-	  * @return First inserted paths, then paths that already existed in the database.
-	  * Both are grouped by their domain ids.
-	  */
-	def store(paths: Map[Int, Set[String]])(implicit connection: Connection) = {
-		if (paths.nonEmpty) {
-			// Finds existing paths under the specified domains
-			val existingPerDomain = withinDomains(paths.keys).pull.groupBy { _.domainId }
-			// .view.mapValues { _.map { p => p.path.toLowerCase -> p.id }.toMap }.toMap
-			// Inserts missing paths (case-insensitive)
-			val inserted = model.insert(paths.flatMap { case (domainId, paths) =>
-				val existing = existingPerDomain.getOrElse(domainId, Empty)
-				paths.distinctBy { _.toLowerCase }
-					.filterNot { p => existing.exists { _.path ~== p } }
-					.map { p => RequestPathData(domainId, p) }
-			}.toVector)
-			
-			// Returns a domain-grouped list of request paths
-			// Specifies separate groups for existing and inserted items
-			Pair(inserted.groupBy { _.domainId }, existingPerDomain)
-		}
-		else
-			Pair.twice(Map.empty[Int, Seq[RequestPath]])
-	}
+	// ATTRIBUTES   -----------------------
 	
+	private val storeLock = new AnyRef
+	
+	
+	// OTHER    ---------------------------
+	
+	/**
+	 * Stores the request paths from the specified links
+	 * @param links Links from which request paths are to be stored
+	 * @param connection Implicit DB connection
+	 * @param log Implicit logging implementation for logging warnings about mapping failures
+	 * @return Stored links, where each entry contains 3 values:
+	 *              1. Stored request path, including domain information
+	 *              1. Specified link
+	 *              1. Whether this represents a newly inserted request path
+	 */
+	def storeFromLinks(links: Iterable[Link])(implicit connection: Connection, log: Logger) =
+		storeFrom(links) { _.domain } { _.path } { (path, link, wasInserted) => (path, link, wasInserted) }
 	/**
 	 * Stores a group of request paths
 	 * @param values Values from which request path data is extracted
@@ -69,42 +57,76 @@ object DbRequestPaths
 	def storeFrom[A, R](values: Iterable[A])
 	                   (extractDomain: A => String)(extractPath: A => String)
 	                   (merge: (DetailedRequestPath, A, Boolean) => R)
-	                   (implicit connection: Connection) =
+	                   (implicit connection: Connection, log: Logger) =
 	{
-		// Stores the domains first
-		val domainValues = DbDomains
-			.storeFrom(values)(extractDomain) { (domain, value, wasInserted) => (domain, value, wasInserted) }
-		
-		// Next, stores the request paths
-		val (pathsToInsert, pathsToStore) = domainValues
-			.divideWith { case (domain, value, wasInserted) =>
-				val path = extractPath(value)
-				if (wasInserted) Left((domain, path, value)) else Right((domain, path, value))
+		if (values.isEmpty)
+			Empty
+		else
+			storeLock.synchronized {
+				// Stores the domains first
+				val domainValues = DbDomains
+					.storeFrom(values)(extractDomain) { (domain, value, wasInserted) => (domain, value, wasInserted) }
+				
+				// Next, stores the request paths
+				val (pathsToInsert, pathsToStore) = domainValues
+					.divideWith { case (domain, value, wasInserted) =>
+						val path = extractPath(value)
+						if (wasInserted) Left((domain, path, value)) else Right((domain, path, value))
+					}
+				// Paths on new domains may be inserted without duplicate checks
+				val insertResults = model.insertFrom(pathsToInsert) {
+					case (domain, path, _) => RequestPathData(domain.id, path) } {
+					case (path, (domain, _, value)) => merge(DetailedRequestPath(path, domain), value, true) }
+				// Paths on existing domains must be stored more carefully
+				val storeResults = Store.storeFrom(pathsToStore) { case (domain, path, _) => domain -> path } {
+					case (path, (_, _, original), wasInserted) => merge(path, original, wasInserted) }
+				
+				insertResults ++ storeResults
 			}
-		// Paths on new domains may be inserted without duplicate checks
-		val insertedPaths = model.insertFrom(pathsToInsert) {
-			case (domain, path, _) => RequestPathData(domain.id, path) } {
-			case (path, (domain, _, value)) => (DetailedRequestPath(path, domain), value, true) }
-		// Paths on existing domains must be stored more carefully
-		val storedPathMap = store(
-			pathsToStore.view.map { case (domain: Domain, path: String, _) => domain.id -> path }.toSet[(Int, String)]
-				.groupMap { _._1 } { _._2 })
-			.zipWithSide
-			.map { case (paths, side) =>
-				val wasInserted = side == First
-				paths.view.mapValues { _.view.map { p => p.path.toLowerCase -> (p, wasInserted) }.toMap }.toMap
-			}
-			.merge { _.mergeWith(_) { _ ++ _ } }
-		val storedPaths = pathsToStore.flatMap { case (domain, pathStr, value) =>
-			storedPathMap.get(domain.id).flatMap { pathsMap =>
-				pathsMap.get(pathStr.toLowerCase).map { case (path, wasInserted) =>
-					(DetailedRequestPath(path, domain), value, wasInserted)
-				}
+	}
+	/**
+	  * Stores the specified request paths to the database.
+	  * Avoids inserting duplicates
+	  * @param paths Paths to store. Each entry consists of a domain id, plus the path part as a string.
+	  * @param connection Implicit DB connection
+	  * @return A map (view) which contains the specified paths as keys and detailed request paths as values.
+	 *         Each value is coupled with a boolean that's set to true if it represents a newly inserted entity.
+	  */
+	def store(paths: Set[(Domain, String)])(implicit connection: Connection) =
+		Store.store(paths)
+	
+	
+	// NESTED   ----------------------------
+	
+	private object Store extends CachingVolatileMapStore[(Domain, String), (Int, String), DetailedRequestPath]
+	{
+		override protected def standardize(value: (Domain, String)): (Int, String) = value._1.id -> value._2.toLowerCase
+		override protected def diff(proposed: Set[(Domain, String)], existing: Set[(Int, String)]): Set[(Domain, String)] = {
+			val existingByDomain = existing.groupMap { _._1 } { _._2 }
+			proposed.filterNot { case (domain, path) =>
+				existingByDomain.get(domain.id).exists { _.contains(path.toLowerCase) }
 			}
 		}
 		
-		// Merges the inserted paths with the original values
-		OptimizedIndexedSeq
-			.from(View.concat(insertedPaths, storedPaths).map { case (p, v, inserted) => merge(p, v, inserted) })
+		override protected def pullMatchMap(values: Set[(Domain, String)])
+		                                   (implicit connection: Connection): Map[(Int, String), DetailedRequestPath] =
+		{
+			val domainMap = values.view.map { case (domain, _) => domain.id -> domain }.toMap
+			withinDomains(values.map { _._1.id }).pull.view
+				.map { p =>
+					val domain = domainMap(p.domainId)
+					(domain.id -> p.path.toLowerCase) -> domain(p)
+				}
+				.toMap
+		}
+		
+		override protected def insertAndMap(values: Seq[(Domain, String)])
+		                                   (implicit connection: Connection): Map[(Int, String), DetailedRequestPath] =
+		{
+			// Inserts the request paths and forms the final map
+			model.insertFrom(values) { case (domain, path) => RequestPathData(domain.id, path) } {
+				case (path, (domain, pathStr)) => (domain.id -> pathStr.toLowerCase) -> domain(path) }
+				.toMap
+		}
 	}
 }